1.1.3.9 Beta

+++ scripts/_xkeen/03_delete/06_delete_tmp.sh
@@ -12,7 +12,7 @@ delete_tmp() {
         rm "$register_dir/new_entry.txt"
     fi
 
-    if ! pidof xray >/dev/null || ! pidof mihomo >/dev/null ; then
+    if ! pidof xray >/dev/null && ! pidof mihomo >/dev/null ; then
         if [ -f "/opt/etc/ndm/netfilter.d/proxy.sh" ]; then
             rm "/opt/etc/ndm/netfilter.d/proxy.sh"
         fi
